feat: move to queryparameter
