+++ search-service/src/main/kotlin/com/egm/stellio/search/entity/service/LinkedEntityService.kt
@@ -7,11 +7,11 @@ import com.egm.stellio.search.entity.model.EntitiesQuery
 import com.egm.stellio.search.entity.model.EntitiesQueryFromGet
 import com.egm.stellio.shared.model.APIException
 import com.egm.stellio.shared.model.CompactedEntity
-import com.egm.stellio.shared.model.LinkedEntityQuery
-import com.egm.stellio.shared.model.LinkedEntityQuery.Companion.JoinType
-import com.egm.stellio.shared.model.PaginationQuery
 import com.egm.stellio.shared.model.getRelationshipsObjects
 import com.egm.stellio.shared.model.inlineLinkedEntities
+import com.egm.stellio.shared.queryparameter.LinkedEntityQuery
+import com.egm.stellio.shared.queryparameter.LinkedEntityQuery.Companion.JoinType
+import com.egm.stellio.shared.queryparameter.PaginationQuery
 import com.egm.stellio.shared.util.JsonLdUtils.JSONLD_CONTEXT
 import com.egm.stellio.shared.util.JsonLdUtils.JSONLD_ID_TERM
 import com.egm.stellio.shared.util.JsonLdUtils.compactEntities
