+++ search-service/src/test/kotlin/com/egm/stellio/search/entity/service/LinkedEntityServiceTests.kt
@@ -4,9 +4,9 @@ import arrow.core.right
 import com.egm.stellio.search.entity.model.EntitiesQueryFromGet
 import com.egm.stellio.shared.model.CompactedEntity
 import com.egm.stellio.shared.model.ExpandedEntity
-import com.egm.stellio.shared.model.LinkedEntityQuery
-import com.egm.stellio.shared.model.LinkedEntityQuery.Companion.JoinType
-import com.egm.stellio.shared.model.PaginationQuery
+import com.egm.stellio.shared.queryparameter.LinkedEntityQuery
+import com.egm.stellio.shared.queryparameter.LinkedEntityQuery.Companion.JoinType
+import com.egm.stellio.shared.queryparameter.PaginationQuery
 import com.egm.stellio.shared.util.JsonUtils.deserializeAsMap
 import com.egm.stellio.shared.util.JsonUtils.serializeObject
 import com.egm.stellio.shared.util.LINKED_ENTITY_COMPACT_TYPE
